refactoring event data processes (WIP)
