+++ src/lava/proc/max_pooling/process.py
@@ -48,19 +48,9 @@ class MaxPooling(AbstractProcess):
         self.stride = Var(shape=(2,), init=stride)
 
     @staticmethod
-    def _validate_shape_in(shape_in):
-        if not (len(shape_in) == 2 or len(shape_in) == 3):
-            raise ValueError(f"shape_in should be 2 or 3 dimensional. "
-                             f"{shape_in} given.")
-
-        if not isinstance(shape_in[0], int) or not isinstance(shape_in[1], int):
-            raise ValueError(f"Width and height of shape_in should be integers."
-                             f"{shape_in} given.")
-        if len(shape_in) == 3:
-            if shape_in[2] != 2:
-                raise ValueError(f"Third dimension of shape_in should be "
-                                 f"equal to 2. "
-                                 f"{shape_in} given.")
+    def _validate_shape_in(shape_in: ty.Tuple[int, int, int]):
+        if not len(shape_in) == 3:
+            raise ValueError(f"shape_in should be 3 dimensional. {shape_in} given.")
 
         if shape_in[0] <= 0 or shape_in[1] <= 0:
             raise ValueError(f"Width and height of shape_in should be positive."
